+++ R/ggpie.R
@@ -71,7 +71,7 @@ NULL
 ggpie <- function(data, x, label = NULL, lab.pos = c("out", "in"),
                   lab.font = c(4, "bold", "black"),
                       color = "black", fill = "white", palette = NULL,
-                      size = 1, ggtheme = theme_pubr(),
+                      size = 1, ggtheme = theme_classic2(),
                       ...)
 {