+++ R/ggviolin.R
@@ -91,7 +91,7 @@ ggviolin <- function(data, x, y,
                       select = NULL, order = NULL,
                       add = "mean_se", add.params = list(),
                       error.plot = "pointrange",
-                      ggtheme = theme_pubr(),
+                      ggtheme = theme_classic2(),
                      ...)
 {