+++ R/ggdotchart.R
@@ -55,7 +55,7 @@ NULL
 #' @export
 ggdotchart <- function(data, x, label, group = NULL,
                       color = "black",  palette = NULL,
-                      shape = 19, size = 1,
+                      shape = 19, size = NULL,
                       sorting = c("descending", "ascending"),
                       orientation = c("vertical", "horizontal"),
                       ggtheme = theme_bw(),