+++ R/ggmaplot.R
@@ -59,7 +59,7 @@ NULL
 #'    ggtheme = ggplot2::theme_minimal())
 #' @export
 ggmaplot <- function (data, fdr = 0.05, fc = 1.5, genenames = NULL,
-                     detection_call = NULL, size = 0.9,
+                     detection_call = NULL, size = NULL,
                      font.label = c(12, "plain", "black"), label.rectangle = FALSE,
                      palette = c("#B31B21", "#1465AC", "darkgray"),
                      top = 15, select.top.method = c("padj", "fc"),