+++ NEWS.md
@@ -1,4 +1,16 @@
 
+# ggpubr 0.1.1.999
+   
+  
+## New features
+   
+    
+## Minor changes
+
+## Bug fixed
+
+   
+   
 # ggpubr 0.1.1